+++ templates/gallery/CHANGELOG.md
@@ -1,5 +1,19 @@
 # notro-gallery
 
+## 0.0.11
+
+### Patch Changes
+
+- Updated dependencies [[`5deac91`](https://github.com/mosugi/notro/commit/5deac91cc36708103f6b675d791340a6921e6934)]:
+  - notro-loader@0.2.0
+
+## 0.0.10
+
+### Patch Changes
+
+- Updated dependencies [[`082d14f`](https://github.com/mosugi/notro/commit/082d14f34b944560c3eacce9aa40c949d259750d)]:
+  - notro-loader@0.1.0
+
 ## 0.0.9
 
 ### Patch Changes